feat: WebView 재사용을 위한 컴포넌트화 및 관련 코드 변경
